backend/requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
sqlalchemy
requests
groq
pydantic
psycopg2-binary
